Add Bazel build file implementation
This adds bazel build files for iOS. This includes several files in platform/ios, platform/darwin, platform/default and root directories. There is a build script in platform/ios/platform/ios/scripts that can be used to build the SDK with options for debug builds, release builds and static and dynamic builds. This (I believe) is the only way to statically build the iOS xcframework.