Remove info plist variables
This removes two variables from the info plist files and their ultimate usage within the map. From what I can tell the MGLSemanticVersionString is passed to the http server, which is likely some kind of back end implementation for Mapbox to see the SDK build specifics. I'm not sure if any one else's backend needs or uses this information?

The other part is you can see the MGLSemanticVersionString and not just the regular version in the Map. Which again seems sort of unecessary.

I could not find any usage of MGLCommitHash anywhere so it seems to be outright dead code.

If people want to maintain the MGLSemanticVersionString I think that's okay, it just seems to be of limited and possibly proprietary use.