Simplest possible builder for MapLibre GL Core on pull request
Added a simple Github Workflow that will be triggered if any MapLibre GL Core
path is changed. It will ignore all paths for any platform specific code.

test: Have tested with my local fork for include and src directory.

Next change is to deduce how we can export sources and binaries from this build that
Android build can use. Or find a way for Android to rely on core build.

Fetching ccache with hendrikmuhs/ccache-action@v1.2. We initiate
a single key here that is tied to the github head_ref of the current
code.

We cache Github action ccache folder in ccache-maplibre-core cache.
Only the existing github head_ref is restored.

test: Testing on my own Github fork. I re-ran the same workflow twice
ensure ccache is working.

Sample run:
https://github.com/thehoneymad/maplibre-gl-native/actions/runs/3934620060/jobs/6729568322