Add justfile to simplify docker usage
This adds a [justfile](https://just.systems/man/en/) to simplify docker usage. Eventually more recipes can be added to it to make executing any kind of long complex commands simpler.

Just is a very popular task runner, with 19k stars and 7.5 million downloads.

This is an **optional** feature, and does not have to be used. I think we may want to use this method to keep all CI scripts in one place as well - in order to be able to run them both locally or in a workflow.