modify the transform implementation to allow for concurrent animations
The previous implementation of the transform class overrides the
current animation with the new one. This new implmentation changes
that to allow for concurrent transformations to occur for different
types of transfomrations at the same time. For example, you can
have two concurrent animations where one modifies the zoom
while the other modifies the position of the camera without overriding
the other.

This change is necessary to allow for concurrent camera animations
during navigation.