add the ability to have ios camera animation durations dynamic during navigation
This change makes it so that camera animations during navigations are not
always hard coded to a duration of 1 second. We now replicate what we
do for the user location indicator by measuring the difference between
the last location's timestamp and the current location update
and animate the camera based on that.

This functionality is exposed through a feature flag to maintain
the existing functionality for those that need it.