Bug: Don't abort on potential HTTP/2 connections
In recent Qt6 releases HTTP2 is enabled by default, this was not
the case for Qt5. With this change cancelling a connection via
reply->abort() breaks the protocol. We should not do that.

In addition, if the protocol is broken in an error situation, we
could not recover from this.

Fix #3644